Change Id Int to Long

+++ src/CoreDataStore.Data/Infrastructure/EntityBaseRepository.cs
@@ -59,7 +59,7 @@ namespace CoreDataStore.Data.Infrastructure
             return await query.ToListAsync();
         }
 
-        public T GetSingle(int id)
+        public T GetSingle(long id)
         {
             return _context.Set<T>().FirstOrDefault(x => x.Id == id);
         }
@@ -80,7 +80,7 @@ namespace CoreDataStore.Data.Infrastructure
             return query.Where(predicate).FirstOrDefault();
         }
 
-        public async Task<T> GetSingleAsync(int id)
+        public async Task<T> GetSingleAsync(long id)
         {
             return await _context.Set<T>().FirstOrDefaultAsync(e => e.Id == id);
         }
